+++ besu/src/main/java/org/hyperledger/besu/services/TraceServiceImpl.java
@@ -173,7 +173,9 @@ public class TraceServiceImpl implements TraceService {
                               .map(parent -> calculateExcessBlobGasForParent(protocolSpec, parent))
                               .orElse(BlobGas.ZERO));
 
-              tracer.traceStartTransaction(transaction);
+              final WorldUpdater worldUpdater = chainUpdater.getNextUpdater();
+
+              tracer.traceStartTransaction(worldUpdater, transaction);
 
               final TransactionProcessingResult result =
                   transactionProcessor.processTransaction(
@@ -188,7 +190,7 @@ public class TraceServiceImpl implements TraceService {
                       blobGasPrice);
 
               long transactionGasUsed = transaction.getGasLimit() - result.getGasRemaining();
-              tracer.traceEndTransaction(result.getOutput(), transactionGasUsed, 0);
+              tracer.traceEndTransaction(worldUpdater, result.getOutput(), transactionGasUsed, 0);
 
               results.add(result);
             });