+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 23.7.3
 ### Additions and Improvements
+- Add access to an immutable world view to start/end transaction hooks in the tracing API 
 
 ### Breaking Changes
 - Removed support for Kotti network (ETC) [#5816](https://github.com/hyperledger/besu/pull/5816)